Adding profiling files examples and instructions to run in Readme
